Update integration tests to use new base class
